requirements.txt
# Identity Reconciliation API Dependencies
# Core dependencies for FastAPI identity reconciliation system
# Compatible with Python 3.7+

# Web framework
fastapi==0.68.0
uvicorn==0.15.0

# Database
sqlalchemy==1.4.23
psycopg2-binary==2.9.1

# Data validation
pydantic==1.8.2

# Environment configuration
python-dotenv==0.19.0

# AWS Lambda support (for later deployment)
mangum==0.12.3

# Development dependencies
pytest==6.2.5
pytest-asyncio==0.15.1 